added condition to remove kernel modules

+++ rpi_apt_clean.sh
@@ -10,9 +10,16 @@ sudo rm -rf /usr/share/man/
 sudo rm -rf /usr/share/locale/
 sudo rm -rf /etc/sgml/
 #make sure to do an upgrade to 4.19 using rpi-upgrade, or else you just bricked.
-sudo rm -rf /lib/modules/4.14.79-v7+/
-sudo rm -rf /lib/modules/4.14.*
-sudo rm -rf /lib/modules/4.19.32+
+kerver = $(uname -r)
+if ["$kerver" != "4.14.79-v7+"]; then
+  sudo rm -rf /lib/modules/4.14.79-v7+/
+fi
+if ["$kerver" != "4.14.79+"]; then
+  sudo rm -rf /lib/modules/4.14.79+
+fi
+if ["$kerver" != "4.19.32+"]; then
+  sudo rm -rf /lib/modules/4.19.32+
+fi
 sudo rm -rf /boot.bak
 sudo rm -rf /var/lib/apt/lists
 sudo rm -r  /var/log/*
